+++ src/cst.h
@@ -211,8 +211,8 @@ public:
 
 class Real: public Node {
 public:
-	double value = 0.0;
-	Real(double value, Type* ty);
+	long double value = 0.0L;
+	Real(long double value, Type* ty);
 	const char* diagnostic_kind() const override;
 	ConstEvalResult const_eval(ConstEvalContext& ctx) const override;
 	void print_diagnostic_definition(ErrorLetContext* ctx, std::ostringstream& out, unsigned indent) const override;
